feat: save-moment 기능 구현

+++ data/src/main/java/com/wakeup/data/database/entity/MomentPictureEntity.kt
@@ -24,6 +24,6 @@ import androidx.room.PrimaryKey
 )
 data class MomentPictureEntity(
     @PrimaryKey(autoGenerate = true) val id: Long = 0L,
-    @ColumnInfo(name = "moment_id", index = true) val momentId: Int,
-    @ColumnInfo(name = "picture_id", index = true) val pictureId: Int,
+    @ColumnInfo(name = "moment_id", index = true) val momentId: Long,
+    @ColumnInfo(name = "picture_id", index = true) val pictureId: Long,
 )
